last typehint check + tests correction
